--- KL25Z Code/Headers/motors.h
@@ -1,19 +1,0 @@
-#ifndef MOTORS_HEADER
-#define MOTORS_HEADER
-
-#include "foundation.h"
-
-/**
-* TPM2_CH0 -> Rear-Left forward
-* TPM2_CH1 -> Rear-Left reverse
-* TPM0_CH4 -> Front-Left forward
-* TPM0_CH5 -> Front-Left reverse
-* TPM0_CH0 -> Rear-Right forward
-* TPM0_CH2 -> Rear-Right reverse
-* TPM0_CH3 -> Front-Right forward
-* TPM0_CH1 -> Front-Right reverse
-*/
-
-void stop(void);
-
-#endif
