+++ KL25Z Code/Objects/ExtDll.iex
@@ -1,0 +1,2 @@
+[EXTDLL]
+Count=0
