+++ KL25Z Code/Objects/RobotTest.build_log.htm
@@ -21,59 +21,21 @@ Target DLL:      CMSIS_AGDI.dll V1.32.0.0
 Dialog DLL:      TARMCM1.DLL V1.14.1.0
  
 <h2>Project:</h2>
-C:\Users\Jonathan\Documents\CG2271 stuff\Boards\NXP\FRDM-KL25Z\RTOS\KL25Z Code\RtosProject.uvprojx
+C:\Users\Jonathan\Documents\CG2271 stuff\Boards\NXP\FRDM-KL25Z\RobotTest\RobotTest.uvprojx
 Project File Date:  10/17/2021
 
 <h2>Output:</h2>
 *** Using Compiler 'V5.06 update 6 (build 750)', folder: 'C:\Keil_v5\ARM\ARMCC\Bin'
-Rebuild target 'Target 1'
-assembling startup_MKL25Z4.s...
-compiling RTX_Config.c...
-compiling rtx_lib.c...
-compiling motors.c...
-Headers\foundation.h(16): warning:  #1-D: last line of file ends without a newline
-  #endif
-Headers\motors.h(19): warning:  #1-D: last line of file ends without a newline
-  #endif
-Headers\motors.c(12): warning:  #1-D: last line of file ends without a newline
-  }
-Headers\motors.c: 3 warnings, 0 errors
-compiling UART.c...
-Headers\foundation.h(16): warning:  #1-D: last line of file ends without a newline
-  #endif
-Headers\UART.h(16): warning:  #1-D: last line of file ends without a newline
-  #endif
-Headers\UART.c(42): warning:  #188-D: enumerated type mixed with another type
-  			dir = (serialValue & 0b11101111);
-Headers\UART.c(47): warning:  #1-D: last line of file ends without a newline
-  }
-Headers\UART.c: 4 warnings, 0 errors
-compiling PWM.c...
-Headers\foundation.h(16): warning:  #1-D: last line of file ends without a newline
-  #endif
-Headers\PWM.h(22): warning:  #1-D: last line of file ends without a newline
-  #endif
-Headers\PWM.c(92): warning:  #1-D: last line of file ends without a newline
-  }
-Headers\PWM.c: 3 warnings, 0 errors
-compiling Main.c...
-.\Headers\foundation.h(16): warning:  #1-D: last line of file ends without a newline
-  #endif
-.\Headers\UART.h(16): warning:  #1-D: last line of file ends without a newline
-  #endif
-.\Headers\PWM.h(22): warning:  #1-D: last line of file ends without a newline
-  #endif
-.\Headers\motors.h(19): warning:  #1-D: last line of file ends without a newline
-  #endif
-Main.c: 4 warnings, 0 errors
-compiling system_MKL25Z4.c...
+Build target 'Target 1'
+compiling main.c...
+constants.h(12): warning:  #1-D: last line of file ends without a newline
+  } tasks;
+main.c(89): warning:  #188-D: enumerated type mixed with another type
+  			dir = (serialValue & 0b11101111);		
+main.c: 2 warnings, 0 errors
 linking...
-.\Objects\RtosProject.axf: Error: L6218E: Undefined symbol mainSem (referred from main.o).
-.\Objects\RtosProject.axf: Error: L6218E: Undefined symbol moveSem (referred from main.o).
-Not enough information to list image symbols.
-Not enough information to list load addresses in the image map.
-Finished: 2 information, 0 warning and 2 error messages.
-".\Objects\RtosProject.axf" - 2 Error(s), 14 Warning(s).
+Program Size: Code=9728 RO-data=500 RW-data=5132 ZI-data=356  
+".\Objects\RobotTest.axf" - 0 Error(s), 2 Warning(s).
 
 <h2>Software Packages used:</h2>
 
@@ -108,26 +70,25 @@ Package Vendor: Keil
    * Component: ARM::CMSIS:CORE:5.3.0
 
    * Component: ARM::CMSIS:RTOS2:Keil RTX5:Library:5.5.1
-      Source file:   CMSIS\RTOS2\RTX\Config\RTX_Config.c
       Include file:  CMSIS\RTOS2\RTX\Include\rtx_os.h
       Include file:  CMSIS\RTOS2\RTX\Config\RTX_Config.h
+      Source file:   CMSIS\RTOS2\RTX\Config\RTX_Config.c
       Source file:   CMSIS\RTOS2\RTX\Template\main.c
-      Source file:   CMSIS\RTOS2\RTX\Template\Events.c
-      Source file:   CMSIS\RTOS2\RTX\Template\MemPool.c
-      Source file:   CMSIS\RTOS2\RTX\Source\rtx_lib.c
+      Source file:   CMSIS\RTOS2\RTX\Template\Timer.c
       Source file:   CMSIS\RTOS2\RTX\Template\Thread.c
-      Library file:  CMSIS\RTOS2\RTX\Library\ARM\RTX_CM0.lib
       Source file:   CMSIS\RTOS2\RTX\Template\svc_user.c
       Source file:   CMSIS\RTOS2\RTX\Template\MsgQueue.c
-      Source file:   CMSIS\RTOS2\RTX\Template\Semaphore.c
+      Source file:   CMSIS\RTOS2\RTX\Template\Events.c
+      Library file:  CMSIS\RTOS2\RTX\Library\ARM\RTX_CM0.lib
+      Source file:   CMSIS\RTOS2\RTX\Source\rtx_lib.c
+      Source file:   CMSIS\RTOS2\RTX\Template\MemPool.c
       Source file:   CMSIS\RTOS2\RTX\Template\Mutex.c
-      Source file:   CMSIS\RTOS2\RTX\Template\Timer.c
+      Source file:   CMSIS\RTOS2\RTX\Template\Semaphore.c
 
    * Component: Keil::Device:Startup:2.5.0
-      Source file:   Device\Source\system_MKL25Z4.c
-      Source file:   Device\Source\ARM\startup_MKL25Z4.s
       Include file:  Device\Source\system_MKL25Z4.h
-Target not created.
+      Source file:   Device\Source\ARM\startup_MKL25Z4.s
+      Source file:   Device\Source\system_MKL25Z4.c
 Build Time Elapsed:  00:00:01
 </pre>
 </body>
