+++ KL25Z Code/Objects/RobotTest.lnp
@@ -1,13 +1,10 @@
 --cpu Cortex-M0+
 ".\objects\main.o"
-".\objects\motors.o"
-".\objects\pwm.o"
-".\objects\uart.o"
-"..\..\..\..\..\..\..\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\RTOS2\RTX\Library\ARM\RTX_CM0.lib"
+"..\..\..\..\..\..\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\RTOS2\RTX\Library\ARM\RTX_CM0.lib"
 ".\objects\rtx_lib.o"
 ".\objects\rtx_config.o"
 ".\objects\startup_mkl25z4.o"
 ".\objects\system_mkl25z4.o"
 --ro-base 0x00000000 --entry 0x00000000 --rw-base 0x1FFFF000 --entry Reset_Handler --first __Vectors --strict --summary_stderr --info summarysizes --map --load_addr_map_info --xref --callgraph --symbols
 --info sizes --info totals --info unused --info veneers
---list ".\Listings\RtosProject.map" -o .\Objects\RtosProject.axf
+--list ".\Listings\RobotTest.map" -o .\Objects\RobotTest.axf
