--- KL25Z Code/Objects/motors.d
@@ -1,15 +1,0 @@
-.\objects\motors.o: Headers\motors.c
-.\objects\motors.o: Headers\motors.h
-.\objects\motors.o: Headers\foundation.h
-.\objects\motors.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include\MKL25Z4.h
-.\objects\motors.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
-.\objects\motors.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\core_cm0plus.h
-.\objects\motors.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_version.h
-.\objects\motors.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_compiler.h
-.\objects\motors.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_armcc.h
-.\objects\motors.o: .\RTE\Device\MKL25Z128xxx4\system_MKL25Z4.h
-.\objects\motors.o: .\RTE\_Target_1\RTE_Components.h
-.\objects\motors.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include\MKL25Z4.h
-.\objects\motors.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\RTOS2\Include\cmsis_os2.h
-.\objects\motors.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stddef.h
-.\objects\motors.o: Headers\constants.h
