--- KL25Z Code/Objects/uart.d
@@ -1,16 +1,0 @@
-.\objects\uart.o: Headers\UART.c
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include\MKL25Z4.h
-.\objects\uart.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\core_cm0plus.h
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_version.h
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_compiler.h
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_armcc.h
-.\objects\uart.o: .\RTE\Device\MKL25Z128xxx4\system_MKL25Z4.h
-.\objects\uart.o: Headers\UART.h
-.\objects\uart.o: Headers\foundation.h
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include\MKL25Z4.h
-.\objects\uart.o: .\RTE\_Target_1\RTE_Components.h
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include\MKL25Z4.h
-.\objects\uart.o: C:\Users\Jonathan\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\RTOS2\Include\cmsis_os2.h
-.\objects\uart.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stddef.h
-.\objects\uart.o: Headers\constants.h
