+++ KL25Z Code/RTE/_Target_1/RTE_Components.h
@@ -3,7 +3,7 @@
  * Auto generated Run-Time-Environment Configuration File
  *      *** Do not modify ! ***
  *
- * Project: 'RtosProject' 
+ * Project: 'RobotTest' 
  * Target:  'Target 1' 
  */
